requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
diskcache==5.6.3
faiss-cpu==1.7.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
llama_cpp_python==0.2.44
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==4.0.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
streamlit==1.31.1
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.2
transformers==4.37.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
yarl==1.9.4
zipp==3.17.0
